refactor(auth): split agent capability execution
- move delegated capability execution and error mapping into agent/execution
- move agent runner service wiring into agent/runner
- keep agent/options as the Better Auth option surface with compatible re-exports
